libagl: eglSwapInterval fix
SW egl returned always EGL_FALSE even when the display was valid.
This made calls to eglSwapInterval to fail on HW since the wrapper
fails if either SW or HW egl fails.

Letting SW eglSwapInterval to return allways true when display is
valid is a-ok since spec says that interval parameter is silently
clamped.

Change-Id: Icd454da3db2982ffe19b62873823fcd12543f690